@@ -1,7 +1,7 @@
 {
   "name": "raw-body",
   "description": "Get and validate the raw body of a readable stream.",
-  "version": "2.5.2",
+  "version": "3.0.1",
   "author": "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
@@ -12,25 +12,25 @@
   "dependencies": {
     "bytes": "3.1.2",
     "http-errors": "2.0.0",
-    "iconv-lite": "0.4.24",
+    "iconv-lite": "0.7.0",
     "unpipe": "1.0.0"
   },
   "devDependencies": {
     "bluebird": "3.7.2",
-    "eslint": "8.34.0",
+    "eslint": "8.57.0",
     "eslint-config-standard": "15.0.1",
-    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-import": "2.29.1",
     "eslint-plugin-markdown": "3.0.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-promise": "6.6.0",
     "eslint-plugin-standard": "4.1.0",
-    "mocha": "10.2.0",
-    "nyc": "15.1.0",
+    "mocha": "10.7.0",
+    "nyc": "17.0.0",
     "readable-stream": "2.3.7",
     "safe-buffer": "5.2.1"
   },
   "engines": {
-    "node": ">= 0.8"
+    "node": ">= 0.10"
   },
   "files": [
     "HISTORY.md",
@@ -44,6 +44,7 @@
     "lint": "eslint .",
     "test": "mocha --trace-deprecation --reporter spec --bail --check-leaks test/",
     "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
-    "test-cov": "nyc --reporter=html --reporter=text npm test"
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
   }
 }
