@@ -3,11 +3,14 @@ import cookieParser from "cookie-parser";
 import logger from "morgan";
 import cors from "cors";
 
-// var createError = require("http-errors");
-// var path = require("path");
-
-// var indexRouter = require("./routes/index");
-// var usersRouter = require("./routes/users");
+import statusRouter from "./routes/status";
+// import authRouter from './routes/auth'
+import userRouter from "./routes/user";
+import challengeRouter from "./routes/challenge";
+import proposalRouter from "./routes/proposal";
+import messageRouter from "./routes/message";
+// import authentication from './middlewares/authentication'
+// import authorization from './middlewares/authorization'
 
 const app = express();
 
@@ -18,6 +21,18 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 // app.use(authorization)
 
+// Rutas principales
+app.use("/", statusRouter);             // GET / y /status
+app.use("/users", userRouter);          // GET /users?role=empresa
+app.use("/challenges", challengeRouter); // GET /challenges?estado=activo&empresaId=123
+app.use("/proposals", proposalRouter);   // GET /proposals?estado=en%20revision&emprendedorId=456
+app.use("/messages", messageRouter);     // GET /messages?empresaId=...&desafioId=...
+
+// Middleware de manejo de errores
+app.use((err: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
+  console.error("❌ Error:", err.message || err);
+  res.status(500).send("Internal server error");
+});
 // app.use('/', statusRouter)
 // app.use('/auth', authRouter)
 // app.use('/users', authentication, userRouter)
